refactor gitrepo and policycheck
