+++ tools/build-tools/src/bumpVersion/createBranch.ts
@@ -7,7 +7,8 @@ import { Context, VersionBumpType } from "./context";
 import { getReleasedPrereleaseDependencies } from "./bumpDependencies";
 import { bumpRepo } from "./bumpVersion";
 import { ReferenceVersionBag, getRepoStateChange } from "./versionBag";
-import { fatal, runPolicyCheckWithFix, } from "./utils";
+import { runPolicyCheckWithFix } from "./policyCheck";
+import { fatal } from "./utils";
 import { MonoRepoKind } from "../common/monoRepo";
 import { Package } from "../common/npmPackage";
 import * as semver from "semver";
